+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Now users can edit the title of a tab.
+
 ## [1.3.1] - 2026-05-09
 
 ### Changed